+++ R/predict_topologies_from_sequences.R
@@ -1,7 +1,4 @@
 #' Run PureseqTM directy on a protein sequence
-#' @param protein_sequence a protein sequence, with
-#'   the amino acids as capitals, for
-#'   example `MEILCEDNTSLSSIPNSL`
 #' @inheritParams default_params_doc
 #' @return a topology as a string of zeroes and ones, where a one denotes
 #' that the corresponding amino acid is located within the membrane.
@@ -15,14 +12,21 @@
 #' }
 #' @author Richèl J.C. Bilderbeek
 #' @export
-predict_topology_from_sequence <- function(
-  protein_sequence,
+predict_topologies_from_sequences <- function(
+  protein_sequences,
   folder_name = get_default_pureseqtm_folder(),
   temp_fasta_filename = tempfile(fileext = ".fasta")
 ) {
-  pureseqtmr::predict_topologies_from_sequences(
-    protein_sequences = protein_sequence,
-    folder_name = folder_name,
-    temp_fasta_filename = temp_fasta_filename
+  pureseqtmr::check_pureseqtm_installation(folder_name)
+  t_sequences <- tibble::tibble(
+    name = paste0("protein_", seq_along(protein_sequences)),
+    sequence = protein_sequences
   )
+  save_tibble_as_fasta_file(
+    t = t_sequences,
+    fasta_filename = temp_fasta_filename
+  )
+  t_topologies <- pureseqtmr::predict_topology(temp_fasta_filename)
+  file.remove(temp_fasta_filename)
+  t_topologies$topology
 }